GlobalMap::render() didn't compile in vs2010.
The compilation failed because initializing an arrays size with a
non constant expression is not valid C++ and not accepted by the
vs2010 compiler. Though it is accepted by gcc due to it supporting
variable sized arrays via a compiler extension.

I changed the code to use an std::vector.